requirements.txt
psycopg2-binary
requests
jsonlib
streamlit
pandas